Move achievements button to top left of homescreen
